backend/requirements.txt
async-timeout==4.0.2
attrs==22.2.0
exceptiongroup==1.1.1
hiredis==2.2.2
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.2
redis==4.5.1
tomli==2.0.1
boto3
